+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/services/configurers/implementations/DefaultDocumentConfigurer.java
@@ -55,7 +55,7 @@ public class DefaultDocumentConfigurer implements DocumentConfigurer<DefaultDocu
         document.setTitle(fileName);  // set the title to the document config
         document.setUrl(documentManager.getDownloadUrl(fileName, true));  // set the URL to download a file to the document config
         document.setUrlUser(documentManager.getFileUri(fileName, false));  // set the file URL to the document config
-        document.setDirectUrl(documentManager.getDownloadUrl(fileName, false));
+        document.setDirectUrl(wrapper.getIsEnableDirectUrl() ? documentManager.getDownloadUrl(fileName, false) : "");
         document.setFileType(fileUtility.getFileExtension(fileName).replace(".",""));  // set the file type to the document config
         document.getInfo().setFavorite(wrapper.getFavorite());  // set the favorite parameter to the document config