+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/services/configurers/wrappers/DefaultFileWrapper.java
@@ -37,4 +37,5 @@ public class DefaultFileWrapper {
     private Action action;
     private String actionData;
     private Boolean canEdit;
+    private Boolean isEnableDirectUrl;
 }